requirements.txt
absl-py
aenum
aiofiles
aiohappyeyeballs
aiohttp
aiosignal
annotated-types
anyio
asgiref
async-timeout
attrs
backoff
bcrypt
bidict
blinker
boto3
botocore
build
CacheControl
cachetools
certifi
cffi
charset-normalizer
chroma-hnswlib
chromadb
click
coloredlogs
contourpy
cryptography
cycler
dataclasses-json
Deprecated
distro
durationpy
exceptiongroup
fastapi
filelock
firebase-admin
Flask
Flask-Cors
Flask-SocketIO
flatbuffers
fonttools
fpdf
frozenlist
fsspec
future
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-cloud-core
google-cloud-firestore
google-cloud-storage
google-crc32c
google-resumable-media
googleapis-common-protos
groq
grpcio
grpcio-status
h11
httpcore
httplib2
httptools
httpx
httpx-sse
huggingface-hub
humanfriendly
idna
importlib_metadata
importlib_resources
itsdangerous
jax
jaxlib
Jinja2
jmespath
joblib
jsonpatch
jsonpointer
kiwisolver
kubernetes
langchain
langchain-community
langchain-core
langchain-groq
langchain-text-splitters
langsmith
line-bot-sdk
livekit
livekit-api
livekit-protocol
lxml
markdown-it-py
MarkupSafe
marshmallow
matplotlib
mdurl
mediapipe
ml_dtypes
mmh3
monotonic
mpmath
msgpack
multidict
mypy-extensions
networkx
numpy
oauthlib
onnxruntime
opencv-contrib-python
opentelemetry-api
opentelemetry-exporter-otlp-proto-common
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-instrumentation
opentelemetry-instrumentation-asgi
opentelemetry-instrumentation-fastapi
opentelemetry-proto
opentelemetry-sdk
opentelemetry-semantic-conventions
opentelemetry-util-http
opt_einsum
orjson
overrides
packaging
pillow
posthog
propcache
proto-plus
protobuf
pyasn1
pyasn1_modules
pycparser
pydantic
pydantic-settings
pydantic_core
pydub
Pygments
PyJWT
pyparsing
PyPika
pyproject_hooks
python-dateutil
python-docx
python-dotenv
python-engineio
python-socketio
PyYAML
regex
requests
requests-oauthlib
requests-toolbelt
rich
rsa
s3transfer
safetensors
scikit-learn
scipy
sentence-transformers
sentencepiece
shellingham
simple-websocket
six
sniffio
sounddevice
SQLAlchemy
starlette
sympy
tenacity
threadpoolctl
tokenizers
tomli
torch
tqdm
transformers
typer
types-protobuf
typing-inspect
typing_extensions
uritemplate
urllib3
uvicorn
uvloop
watchfiles
websocket-client
websockets
Werkzeug
wrapt
wsproto
yarl
zipp
zstandard
